+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     
 ### Added
 - Add browser support information to content share guide.
+- Add support in `MessagingSession` to allow websocket connection for messaging service to enable Prefetch feature.
+
 ### Removed
     
 ### Changed